requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
matplotlib==3.9.0

langchain==0.2.5
langchain-openai==0.1.8
langgraph==0.0.55

openai==1.35.3

python-dotenv==1.0.1
python-multipart==0.0.9

redis==5.0.4
gunicorn==22.0.0